switched heuristic to double

+++ src/main/java/com/furkandolasik/mazepuzzle/solver/InterfaceNode.java
@@ -32,9 +32,9 @@ public interface InterfaceNode {
 
     public void setSelected(boolean selected);
 
-    int getHeuristic();
+    double getHeuristic();
 
-    void setHeuristic(int heuristic);
+    void setHeuristic(double heuristic);
 
     int getCost();
